The simplest way, I presume, is to do a type casting on the variable of integer, so that the compiler can operate with the other float type operand.

+++ rust_book/variables/src/main.rs
@@ -11,6 +11,6 @@ fn main() {
     // Now, let's try operating with variables of different types.
     let x: u8 = 10;
     let w: f32 = 2.5;
-    let r: f32 = x + w;
+    let r: f32 = x as f32 + w;
     println!("{r}");
 }
